Continue with race logic: added acceleration changes, speed and damage

+++ Code/prGame/core/src/com/mygdx/game/model/movement/AIControlledStrategy.java
@@ -33,7 +33,7 @@ public class AIControlledStrategy implements MovementStrategy {
         float laneRightLimit = laneLeftLimit + Lane.WIDTH;
         float movableLeftLimit = movable.getHitbox().x;
         float movableRightLimit = movable.getHitbox().x + movable.getHitbox().width;
-        int speed = movable.getSpeed();
+        float speed = movable.getSpeed();
 
         // Fixed vertical movement
         movable.adjustY(speed * delta);
