+++ Code/prGame/core/src/com/mygdx/game/model/movement/Movable.java
@@ -9,6 +9,6 @@ public interface Movable {
     void setMovementStrategy(MovementStrategy movementStrategy);
     MovementStrategy getMovementStrategy();
     Rectangle getHitbox();
-    int getSpeed();
-    void setSpeed(int speed);
+    float getSpeed();
+    void setSpeed(float speed);
 }
